sec try

+++ file_io/1-create_file.c
@@ -12,7 +12,7 @@ int create_file(const char *filename, char *text_content)
 {
 size_t l = 0;
 int o;
-int w;
+size_t w;
 if (filename == NULL)
 	return (-1);
 o = open(filename, O_CREAT | O_WRONLY | O_TRUNC, 0600);
